refactor: Move config to generator.common.config and update imports
